requirements.txt
flask
pillow
pytesseract
opencv-python
numpy
